refactor: Move share components to dialog subpackage
