+++ identity_presentation/src/androidMain/kotlin/net/thechance/mena/identity/presentation/screen/profile/components/dialog/share/clipEntryOf.android.kt
@@ -1,4 +1,4 @@
-package net.thechance.mena.identity.presentation.screen.profile.components.share
+package net.thechance.mena.identity.presentation.screen.profile.components.dialog.share
 
 import android.content.ClipData
 import androidx.compose.ui.platform.ClipEntry
